backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-agent-assistant"
version = "0.1.0"
description = "High-performance RAG Intelligent Agent Assistant"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Claude", email = "noreply@anthropic.com"}
]

dependencies = [
    # Core framework
    "fastapi[all]>=0.104.0",
    "uvicorn[standard]",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # RAG Stack
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "pymilvus>=2.3.0",

    # Document processing
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "markdown>=3.5.0",
    "tiktoken>=0.5.0",

    # Model services
    "tavily-python>=0.3.0",
    "sentence-transformers>=2.2.0",
    "FlagEmbedding>=1.2.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",

    # Utilities
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
    "pydantic-extra-types>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "ruff>=0.1.5",
    "mypy>=1.6.0",
]

[tool.setuptools.packages.find]
where = ["app"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]
